+++ libs/integration-interface/src/lib/topics/permissions-rpc/v1/permissions-rpc.model.ts
@@ -1,0 +1,5 @@
+export interface PermissionsRpc {
+  appId: string
+  productName: string
+  permissions?: Array<string>
+}