+++ libs/integration-interface/src/lib/topics/permissions/v1/permissions.topic.ts
@@ -1,7 +1,7 @@
-import { SyncableTopic } from "@onecx/accelerator";
+import { SyncableTopic } from '@onecx/accelerator'
 
 export class PermissionsTopic extends SyncableTopic<string[]> {
-    constructor() {
-      super('permissions', 1)
-    }
+  constructor() {
+    super('permissions', 1)
   }
+}